Some fixes..
git-svn-id: https://svn.java.net/svn/docsources~svn/trunk@1813 3b25252b-017b-4ae5-8f61-fa87a855a72e

+++ DocSources/src/main/java/org/medici/docsources/controller/community/AjaxController.java
@@ -75,7 +75,7 @@ public class AjaxController {
 		Map<String, Object> model = new HashMap<String, Object>();
 		
 		try{
-			Integer numberOfNewMessages = getCommunityService().checkNewMessages();
+			Long numberOfNewMessages = getCommunityService().checkNewMessages();
 			model.put("numberOfNewMessages", numberOfNewMessages.toString());
 			model.put("numberOfNewMessages", "0");
 			model.put("newMessages", (numberOfNewMessages>0) ? "true" : "false");
